Create README.md
# Robux-Farm-v2
ROBUX FARM V2 IS FINALLY OUT. Please make sure you have your antivirus turned off temporarily. (Real time scanning) Since this is a Roblox exploit, it gets falsely flagged as a virus due to the fact that it is an exploit. There are no real viruses in my exploit, so there is nothing to worry about.